Feat: Lexica linked to AWS dynamodb
The app now pull the lexica from the AWS dynamodb and only retrieve the latest version. The app also store the pulled data into the cache, however the concistancy of this method has to be reworked as it is not fully exploited. It lacks the capability to reduce the number of requests.